% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{func_POSm4}
\alias{func_POSm4}
\title{Example ODE function: chain model with positive regulation.}
\format{
R file with definition of function func_POSm4 that takes as
input arguments time t (dimension 1), variable values x (dimension 4), and 
kinetic parameter values klin (dimension 8) and knonlin (dimension 2).
}
\source{
The chain model was used in 
Baum K, Politi AZ, Kofahl B, Steuer R, Wolf J. Feedback, Mass 
 Conservation and Reaction Kinetics Impact the Robustness of Cellular 
 Oscillations. PLoS Comput Biol. 2016;12(12):e1005298.
}
\usage{
func_POSm4
}
\description{
The file contains the function definition an ordinary differential equation
model of a chain model of 4 variables with positive feedback.
}
\keyword{datasets}
